[dotnet] Ensure Status endpoint is available to all WebDriver implementations
