+++ dotnet/src/webdriver/WebDriver.cs
@@ -21,8 +21,12 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Globalization;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
 using OpenQA.Selenium.Interactions;
 using OpenQA.Selenium.Internal;
+using Newtonsoft.Json;
 
 namespace OpenQA.Selenium
 {
@@ -77,6 +81,39 @@ namespace OpenQA.Selenium
         public ICapabilities Capabilities
         {
             get { return this.capabilities; }
+        }
+
+        public static Status GetStatus(Uri remoteAddress)
+        {
+            String response = new TaskFactory(CancellationToken.None,
+                    TaskCreationOptions.None,
+                    TaskContinuationOptions.None,
+                    TaskScheduler.Default)
+                .StartNew(() => GetStatusResponse(remoteAddress))
+                .Unwrap()
+                .GetAwaiter()
+                .GetResult();
+
+            Dictionary<string, object> deserializedResponse =
+            JsonConvert.DeserializeObject<Dictionary<string, object>>(response, new ResponseValueJsonConverter());
+
+            Dictionary<string, object> value = (Dictionary<string, object>)deserializedResponse["value"];
+
+            bool ready = (bool)value["ready"];
+            String message = (string)value["message"];
+
+            return new Status(ready, message);
+        }
+
+        private static async Task<String> GetStatusResponse(Uri remoteAddress)
+        {
+            using (HttpClient client = new HttpClient())
+            {
+                HttpRequestMessage request = new HttpRequestMessage(HttpMethod.Get, remoteAddress + "status");
+                HttpResponseMessage responseMessage = await client.SendAsync(request);
+                String responseString = await responseMessage.Content.ReadAsStringAsync();
+                return responseString;
+            }
         }
 
         /// <summary>
